# Fix: GLPK with no constraints