--- a/cvxpy/reductions/solvers/conic_solvers/glpk_mi_conif.py
+++ b/cvxpy/reductions/solvers/conic_solvers/glpk_mi_conif.py
@@ -19,7 +19,6 @@ import cvxpy.settings as s
 from cvxpy.reductions.solution import Solution, failure_solution
 from cvxpy.reductions.solvers.conic_solvers import GLPK
 from cvxpy.reductions.solvers.conic_solvers.conic_solver import ConicSolver
-import numpy as np
 
 
 class GLPK_MI(GLPK):
@@ -53,15 +52,6 @@ class GLPK_MI(GLPK):
         import cvxopt
         import cvxopt.solvers
 
-        # handle missing constraints
-        var_length = data[s.C].size[0]
-        if data[s.A] is None:
-            data[s.A] = intf.sparse2cvxopt(np.zeros((1, var_length)))
-            data[s.B] = intf.dense2cvxopt(np.zeros((1, 1)))
-        if data[s.G] is None:
-            data[s.G] = intf.sparse2cvxopt(np.zeros((1, var_length)))
-            data[s.H] = intf.dense2cvxopt(np.zeros((1, 1)))
-
         # Save original cvxopt solver options.
         old_options = cvxopt.glpk.options.copy()
         # Silence cvxopt if verbose is False.
@@ -69,6 +59,8 @@ class GLPK_MI(GLPK):
             cvxopt.glpk.options["msg_lev"] = "GLP_MSG_ON"
         else:
             cvxopt.glpk.options["msg_lev"] = "GLP_MSG_OFF"
+
+        data = self._handle_missing_constraints(data)
 
         # Apply any user-specific options.
         # Rename max_iters to maxiters.