--- a/cvxpy/tests/solver_test_helpers.py
+++ b/cvxpy/tests/solver_test_helpers.py
@@ -245,6 +245,17 @@ def lp_5() -> SolverTestHelper:
     return sth
 
 
+def lp_6() -> SolverTestHelper:
+    """Test LP with no constraints"""
+    x = cp.Variable()
+    from cvxpy.expressions.constants import Constant
+    objective = cp.Maximize(Constant(0.23) * x)
+    obj_pair = (objective, np.inf)
+    var_pairs = [(x, None)]
+    sth = SolverTestHelper(obj_pair, var_pairs, [])
+    return sth
+
+
 def socp_0() -> SolverTestHelper:
     x = cp.Variable(shape=(2,))
     obj_pair = (cp.Minimize(cp.norm(x, 2) + 1), 1)
@@ -832,6 +843,17 @@ class StandardTestLPs:
     @staticmethod
     def test_lp_5(solver, places: int = 4, duals: bool = True,  **kwargs) -> SolverTestHelper:
         sth = lp_5()
+        sth.solve(solver, **kwargs)
+        sth.verify_objective(places)
+        sth.check_primal_feasibility(places)
+        if duals:
+            sth.check_complementarity(places)
+            sth.check_dual_domains(places)
+        return sth
+
+    @staticmethod
+    def test_lp_6(solver, places: int = 4, duals: bool = True, **kwargs) -> SolverTestHelper:
+        sth = lp_6()
         sth.solve(solver, **kwargs)
         sth.verify_objective(places)
         sth.check_primal_feasibility(places)