--- a/cvxpy/tests/solver_test_helpers.py
+++ b/cvxpy/tests/solver_test_helpers.py
@@ -902,6 +902,7 @@ class StandardTestLPs:
         sth.verify_primal_values(places)
         return sth
 
+
 class StandardTestSOCPs:
 
     @staticmethod