## Description
When no constraints are specified, the GLPK_MI and GLPK solvers would previously fail. Here we convert missing constraint matrices from `None` into trivial matrices that are valid for use by `cvxopt.glpk`.
Issue link (if applicable): https://github.com/cvxpy/cvxpy/issues/1614

### Benchmark before this change:
```
=========================================================================== test session starts ============================================================================
platform linux -- Python 3.8.2, pytest-6.2.5, py-1.11.0, pluggy-1.0.0
rootdir: /home/allen/work/community/cvxpy, configfile: pyproject.toml
collected 11 items                                                                                                                                                         

cvxpy/tests/test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=8.369e-01 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=3.681e+00 s , std=0.000e+00 s (1 iterations)
.least_squares: avg=4.734e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.038715335
cvxpy time:  0.41344025505737303
Quadratic canonicalization
(OSQP) solver time:  0.017661457
cvxpy time:  0.5075722885749512
.qp: avg=9.489e-03 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=3.803e-02 s , std=1.218e-02 s (10 iterations)
.small_lp: avg=1.310e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=1.061e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=1.717e+00 s , std=0.000e+00 s (1 iterations)
.
```

### Benchmark after this change:  (updated to latest commit)
```
=========================================================================== test session starts ============================================================================
platform linux -- Python 3.8.2, pytest-6.2.5, py-1.11.0, pluggy-1.0.0
rootdir: /home/allen/work/community/cvxpy, configfile: pyproject.toml
collected 11 items                                                                                                                                                         

cvxpy/tests/test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=8.326e-01 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=3.606e+00 s , std=0.000e+00 s (1 iterations)
.least_squares: avg=4.615e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.038927703
cvxpy time:  0.40401014795214846
Quadratic canonicalization
(OSQP) solver time:  0.01777715
cvxpy time:  0.5073912688842773
.qp: avg=9.249e-03 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=3.758e-02 s , std=1.457e-02 s (10 iterations)
.small_lp: avg=1.250e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=1.085e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=1.740e+00 s , std=0.000e+00 s (1 iterations)
.
```

## Type of change
- [ ] New feature (backwards compatible)
- [ ] New feature (breaking API changes)
- [x] Bug fix
- [ ] Other (Documentation, CI, ...)

## [Contribution checklist](https://www.cvxpy.org/contributing/index.html#contribution-checklist)
- [x] Add our license to new files.
- [x] Check that your code adheres to our coding style.
- [x] Write unittests.
- [x] Run the unittests and check that they’re passing.
- [x] Run the benchmarks to make sure your change doesn’t introduce a regression.